fix: prevent fatal error when delete person custom field
When we add a custom field to a person and then delete this, could throw
a type error when sum empty string with 1:

```
Fatal error: Uncaught TypeError: Unsupported operand types: string + int i
```

I fixed this adding a cast when fetch a value from $_GET and this value
doesn't exists. The cast need to be equals to 1 because is the lowest
order.

If the column already was deleted (I reproduced this when I made a
refresh after the previous error) and we try to delete again, throws an
error because is trying to delete a non existing column. MySQL have a
condition to only delete the column if exists.

I fixed this:

```
Fatal error: Uncaught mysqli_sql_exception: Can't DROP COLUMN `c4`; check that it exists in /var/www/html/src/Include/Functions.php:175
```

To reproduce before this fix, you will need to add a custom field to a person
and then try to delete this.

Signed-off-by: Vitor Mattos <1079143+vitormattos@users.noreply.github.com>